backend/requirements.txt
fastapi>=0.135.0
uvicorn>=0.42.0
pydantic>=2.12.0
python-dotenv>=1.2.0
requests>=2.33.0
cryptography>=46.0.0
apscheduler>=3.11.0
python-multipart>=0.0.22
aiofiles>=25.1.0
